requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
pillow==10.1.0
openai==1.7.2
PyPDF2==3.0.1
